gogo

+++ src/com/sbs/java/blog/controller/ArticleController.java
@@ -53,38 +53,42 @@ public class ArticleController extends Controller {
 
 		case "replyDelete":
 			return actionReplyDelete();
-			
+
 		case "replyModify":
 			return actionReplyModify();
-			
+
 		case "doReplyModify":
 			return actionDoReplyModify();
 		}
 		return "";
 	}
-	
 
 	private String actionDoReplyModify() {
-		
-		articleService.modifyReply(replyId);
+
+//		articleService.modifyReply(replyId);
 
 		return "html:<script> alert('댓글이 수정되었습니다.'); location.replace('history.back()')</script>";
 	}
 
 	private String actionReplyModify() {
-		int replyId = Util.getInt(req, "replyId");
+		int replyId = Util.getInt(req, "id");
 		
 		return "articleReply/articleReplyModify.jsp";
 	}
 
 	private String actionReplyDelete() {
 		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
-		
-		int replyId = Util.getInt(req, "replyId");
-		
-		articleService.deleteReply(replyId);
 
-		return "html:<script> alert('댓글이 삭제되었습니다.'); location.replace('history.back()')</script>";
+		int replyId = Util.getInt(req, "id");
+		ArticleReply articleReply = articleService.getArticleReplyByReplyId(replyId);
+		int memberId = Integer.parseInt(articleReply.getMemberId());
+		System.out.println(memberId);
+		
+		if (loginedMemberId == memberId) {
+		articleService.deleteReply(replyId);
+			return "html:<script> alert('댓글이 삭제되었습니다.'); location.replace(history.go(-1))</script>";
+		}
+		return "html:<script> alert('권한이 없습니다.'); location.replace(history.back())</script>";
 	}
 
 	private String actionWriteArticleReply() {
@@ -108,8 +112,14 @@ public class ArticleController extends Controller {
 
 	private String actionDelete() {
 		int id = Util.getInt(req, "id");
-		articleService.delete(id);
-		return "html:<script> alert('" + id + "번 게시물이 삭제되었습니다.'); location.replace('list')</script>";
+		Article article = articleService.getArticleById(id);
+		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
+		
+		if (loginedMemberId == article.getId()) {
+			articleService.delete(id);
+			return "html:<script> alert('" + id + "번 게시물이 삭제되었습니다.'); location.replace('list')</script>";
+		}
+		return "html:<script> alert('권한이 없습니다.'); location.replace('../article/detail?id="+article.getId()+"')</script>"; 
 	}
 
 	private String actionDoModify() {
@@ -130,12 +140,15 @@ public class ArticleController extends Controller {
 		if (Util.isNum(req, "id") == false) {
 			return "html:id를 숫자로 입력해 주세요.";
 		}
-		int loginedMemberId = (int) req.getAttribute("loginedMemberId"); 
+		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
 		int id = Util.getInt(req, "id");
-		Article article = articleService.detail(id,loginedMemberId);
+		Article article = articleService.detail(id, loginedMemberId);
 		System.out.println(article);
-		req.setAttribute("article", article);
-		return "article/modify.jsp";
+		if (loginedMemberId == article.getId()) {
+			req.setAttribute("article", article);
+			return "article/modify.jsp";
+		}
+		return "html:<script> alert('권한이 없습니다.'); location.replace('../article/detail?id="+article.getId()+"')</script>";
 	}
 
 	private String actionDoWrite() {
@@ -165,9 +178,9 @@ public class ArticleController extends Controller {
 
 		int id = Util.getInt(req, "id");
 		articleService.increaseHit(id);
-		
+
 		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
-		Article article = articleService.detail(id,loginedMemberId);
+		Article article = articleService.detail(id, loginedMemberId);
 		req.setAttribute("article", article);
 
 		int memberId = article.getMemberId();
@@ -181,9 +194,9 @@ public class ArticleController extends Controller {
 	}
 
 	private String actionList() {
-		
+
 //		long startTime = System.nanoTime();
-		
+
 		int page = 1;
 
 		if (!Util.empty(req, "page") && Util.isNum(req, "page")) {
@@ -228,7 +241,7 @@ public class ArticleController extends Controller {
 				searchKeyword);
 
 		req.setAttribute("articles", articles);
-		
+
 //		long endTime= System.nanoTime();
 //		long estimatedTime = endTime - startTime;
 //		//nano seconds to seconds
